@iancanderson 2022-08-18
The CI failure(s) seem to be happening on `main` too, so unrelated to this PR: https://buildkite.com/rails/rails/builds/88875#0182b103-c9ab-472e-9925-8da478931ecb