@iancanderson 2022-08-18
> Does the performance change if we make this regex match the full set of valid values rather than just this common case?

@jhawthorn it turns out to be very slightly slower, but within error compared to my first implementation. Seems good enough for me! Should I move ahead with the "full set of valid values" regex approach?

```
============================ Common tag name short =============================

Warming up --------------------------------------
     xml_name_escape   159.671k i/100ms
fast_xml_name_escape   343.621k i/100ms
one_regex_xml_name_escape
                       353.869k i/100ms
Calculating -------------------------------------
     xml_name_escape      1.582M (± 3.8%) i/s -      7.984M in   5.055645s
fast_xml_name_escape      3.341M (± 4.7%) i/s -     16.837M in   5.051945s
one_regex_xml_name_escape
                          3.489M (± 4.9%) i/s -     17.693M in   5.085095s

Comparison:
one_regex_xml_name_escape:  3488934.6 i/s
fast_xml_name_escape:  3341149.3 i/s - same-ish: difference falls within error
     xml_name_escape:  1581604.8 i/s - 2.21x  (± 0.00) slower


============================= Common tag name long =============================

Warming up --------------------------------------
     xml_name_escape    76.117k i/100ms
fast_xml_name_escape   266.879k i/100ms
one_regex_xml_name_escape
                       257.459k i/100ms
Calculating -------------------------------------
     xml_name_escape    826.616k (± 4.3%) i/s -      4.186M in   5.074708s
fast_xml_name_escape      2.751M (± 5.0%) i/s -     13.878M in   5.060281s
one_regex_xml_name_escape
                          2.545M (± 5.6%) i/s -     12.873M in   5.077729s

Comparison:
fast_xml_name_escape:  2751129.4 i/s
one_regex_xml_name_escape:  2544673.8 i/s - same-ish: difference falls within error
     xml_name_escape:   826616.2 i/s - 3.33x  (± 0.00) slower


=============================== Custom tag name ================================

Warming up --------------------------------------
     xml_name_escape    78.065k i/100ms
fast_xml_name_escape   242.412k i/100ms
one_regex_xml_name_escape
                       231.727k i/100ms
Calculating -------------------------------------
     xml_name_escape    767.433k (± 6.5%) i/s -      3.825M in   5.012273s
fast_xml_name_escape      2.463M (± 4.3%) i/s -     12.363M in   5.030648s
one_regex_xml_name_escape
                          2.239M (± 6.5%) i/s -     11.355M in   5.095769s

Comparison:
fast_xml_name_escape:  2462694.6 i/s
one_regex_xml_name_escape:  2238947.7 i/s - same-ish: difference falls within error
     xml_name_escape:   767432.9 i/s - 3.21x  (± 0.00) slower


========================== Custom tag name non-ascii ===========================

Warming up --------------------------------------
     xml_name_escape    66.959k i/100ms
fast_xml_name_escape    63.131k i/100ms
one_regex_xml_name_escape
                       206.018k i/100ms
Calculating -------------------------------------
     xml_name_escape    686.496k (± 4.9%) i/s -      3.482M in   5.085224s
fast_xml_name_escape    643.266k (± 4.5%) i/s -      3.220M in   5.016957s
one_regex_xml_name_escape
                          2.010M (± 4.8%) i/s -     10.095M in   5.034582s

Comparison:
one_regex_xml_name_escape:  2010415.6 i/s
     xml_name_escape:   686496.1 i/s - 2.93x  (± 0.00) slower
fast_xml_name_escape:   643266.0 i/s - 3.13x  (± 0.00) slower


================================ Bad tag name 1 ================================

Warming up --------------------------------------
     xml_name_escape    40.491k i/100ms
fast_xml_name_escape    40.465k i/100ms
one_regex_xml_name_escape
                        38.244k i/100ms
Calculating -------------------------------------
     xml_name_escape    393.998k (± 6.7%) i/s -      1.984M in   5.062913s
fast_xml_name_escape    386.906k (± 5.2%) i/s -      1.942M in   5.035445s
one_regex_xml_name_escape
                        377.664k (± 6.6%) i/s -      1.912M in   5.088602s

Comparison:
     xml_name_escape:   393998.2 i/s
fast_xml_name_escape:   386906.4 i/s - same-ish: difference falls within error
one_regex_xml_name_escape:   377663.7 i/s - same-ish: difference falls within error


================================ Bad tag name 2 ================================

Warming up --------------------------------------
     xml_name_escape    15.627k i/100ms
fast_xml_name_escape    15.165k i/100ms
one_regex_xml_name_escape
                        16.028k i/100ms
Calculating -------------------------------------
     xml_name_escape    161.988k (± 5.5%) i/s -    812.604k in   5.032083s
fast_xml_name_escape    161.438k (± 5.7%) i/s -    818.910k in   5.090467s
one_regex_xml_name_escape
                        160.224k (± 5.1%) i/s -    801.400k in   5.016219s

Comparison:
     xml_name_escape:   161987.9 i/s
fast_xml_name_escape:   161437.6 i/s - same-ish: difference falls within error
one_regex_xml_name_escape:   160224.2 i/s - same-ish: difference falls within error
```