@iancanderson 2022-08-18
@jhawthorn https://github.com/rails/rails/pull/45845/commits/8861d08b7b3e30fbef46aba9b14267a3731ec832 adds the "definitely valid regex" you suggested. see comment above for benchmark (the `one_regex_xml_name_escape` case)